使用整个脚本的私有仓库

+++ dd_scripts/shell_default_script.sh
@@ -1,5 +1,58 @@
 #!/bin/sh
-set -e
+
+################################脚本仓库更新/初始化操作 start ################################
+function initNodeEnv() {
+  echo "安装执行脚本需要的nodejs环境及依赖"
+  apk add --update nodejs moreutils npm curl jq
+}
+
+#获取配置的自定义参数,如果有为
+if [ "$1" ]; then
+  initNodeEnv
+  if [ $? -ne 0 ]; then
+    echo "安装执行脚本需要的nodejs环境及依赖出错❌，exit，restart"
+    exit 1
+  else
+    echo "安装执行脚本需要的nodejs环境及依赖成功✅"
+  fi
+  run_cmd=$1
+fi
+
+[ -f /scripts/package.json ] && before_package_json=$(cat /scripts/package.json)
+
+if [ -d /scripts/otherRepo ]; then
+  ddBot -up syncRepo
+fi
+
+if [ $? -ne 0 ]; then
+  echo "更新仓库代码出错❌，跳过"
+else
+  echo "更新仓库代码成功✅"
+fi
+
+if [ ! -d /scripts/node_modules ]; then
+  echo "容器首次启动，执行npm install..."
+  npm install --loglevel error --prefix /scripts
+  if [ $? -ne 0 ]; then
+    echo "npm首次启动安装依赖失败❌，exit，restart"
+    exit 1
+  else
+    echo "npm首次启动安装依赖成功✅"
+  fi
+else
+  if [[ "${before_package_json}" != "$(cat /scripts/package.json)" ]]; then
+    echo "package.json有更新，执行npm install..."
+    npm install --loglevel error --prefix /scripts
+    if [ $? -ne 0 ]; then
+      echo "npackage.json有更新，执行安装依赖失败❌，跳过"
+      exit 1
+    else
+      echo "npackage.json有更新，执行安装依赖成功✅"
+    fi
+  else
+    echo "package.json无变化，跳过npm install..."
+  fi
+fi
 
 if [ -d "/data" ]; then
   if [ -f "/data/env.sh" ]; then
@@ -9,13 +62,16 @@ if [ -d "/data" ]; then
   if [ -d "/data/logs" ]; then
     echo "/data/logs目录已存在，跳过创建。"
   else
+    echo "/data/logs目录不存在，执行创建。"
     mkdir -p /data/logs
   fi
 fi
+################################脚本仓库更新/初始化操作 end ################################
 
-#放弃更新
-echo -e "000" >> /root/.ssh/id_rsa
-cd /scripts && git reset --hard a38137a7defd1a41a5f5438ef8fe0d5becff1982
+if [ ! -d /scripts/otherRepo ]; then
+  echo -e "000" >>/root/.ssh/id_rsa
+  cd /scripts && git reset --hard a38137a7defd1a41a5f5438ef8fe0d5becff1982
+fi
 
 echo "将仓库的docker_entrypoint.sh脚本更新至系统/usr/local/bin/docker_entrypoint.sh内..."
 cat /jds/dd_scripts/docker_entrypoint.sh >/usr/local/bin/docker_entrypoint.sh
@@ -126,13 +182,13 @@ echo "第9步加载最新的定时任务文件..."
 crontab -l >/scripts/befor_cronlist.sh
 
 #自己魔改了企业微信拆分消息通知 暂时没有打包到镜像里面
-if [ -f "/data/sendNotify.js" ];then
+if [ -f "/data/sendNotify.js" ]; then
   cp /data/sendNotify.js /scripts
 fi
 
 echo "增加清理日志，提交互助码到助力池脚本。（测试中）"
 (
-cat <<EOF
+  cat <<EOF
 #!/bin/sh
 set -e
 
@@ -147,7 +203,7 @@ else
     echo "请更新至最新版docker镜像才能自动上传助力码到助力池"
 fi
 EOF
-) > /scripts/submitShareCode.sh
+) >/scripts/submitShareCode.sh
 
 echo "最后加载最新的附加功能定时任务文件..."
 echo "└──替换任务列表的node指令为spnode"
